refactor(worker): optimize async operations
- Replaced sequential async calls with Promise.all for improved performance
  in `CommandExecService`, `StampMetaService`, and `SyncService`.
- Added `start:profile` script and `clinic` dependency in `package.json`
  for profiling application performance.

These changes helped changed the perf profile of the hottest
functions. Prior to this change `compressValue` (related to putting
values in the Redis stream - which is invoked from
`executeSyncActivity`) was the hottest function. And components of
`stampMedata` were serialized and amongst the top 5 hottest
functions.

These functions are occupying lesser time now. In local testing (with
src/dst NFS servers hosted on 2 VMs on the same host as the worker),
SWBuild of 7+ GB (kernel source code for linux and freebsd) now takes
~1 hr, when it was taking more than 1.5 hrs before this change.